requirements.txt
fastapi>=0.101.1,<0.102.0
pydantic>=2.1.1,<2.2.0
uvicorn>=0.23.2,<0.24.0
tortoise-orm>=0.20.0,<0.21.0
aiomysql>=0.2.0,<0.3.0